Rework the beginning of React chapter
I simplified this a lot and jumped straight to Babel. I didn't see much
point in retaining the intermediate parts as you can get straight to the
bleeding edge. Besides 0.13 has been out a while.

This needs to be expanded still. Note that I'm still building on the
same example started earlier. The next step is to port Todo example to
that and discuss basics of React.

Some additional tooling such as eslint and flow could be discussed in
this context as well. eslint helps to understand propTypes and their
value.