from notes to kanban - Simplify `Lane` interface
Now we just pass `lane` object to `Lane`. Easier to understand and less
to explain.

In addition I found a little bug at `deleteNote`. It is better to remove
a note from a Lane before getting rid of the note. This way it won't try
to render something that doesn't exist. It shouldn't show warnings at
console on delete anymore.