react-hot-loader -> react-transform-hmr
This is a big change but this has to be done as react-hot-loader became
deprecated. Now we're relying on a Babel plugin instead.

The Babel plugin approach is nice in that you can now apply arbitrary
transforms on React components easily. So in this way it goes beyond
just HMR.

Closes #248.