Use property initializers across the book
This simplifies the examples somewhat and gets rid of a lot of annoying
binds. I know it's an experimental feature but looks like this is a good
way for now.

Better yet, this will allow proper Flow annotations in the near future
as a new version of Flowtype comes out. That's something which isn't
possible with the constructor based approach.