react and flux - Extract `editing` state from `Editable`
Now it's possible to control `Editable` behavior more easily. If you
want to make the component `editable` by default, simply pass a flag
stating that to it. This can be achieved easily through `create`
actions for example.

In addition I generalized `update` actions to work with arbitrary data.
This solution is more robust than the earlier one and relies on
*object-assign* ponyfill. I tried using `Object.assign` initially, but
then I realized it's not well supported yet.

Closes #282.