Update 02_setting_up_the_project.md
The procedure for removing the current git repository and creating a new one is included twice in this chapter: once as part of the download process, and again as a sidebar.  

It seems better to extract this to a separate section, to make it a single step, which may also be a useful reference for git newbies.  This also makes the later sidebar describing the same action repetitively redundant ;-)